+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Redesign Tools tab with tool inventories, init errors, and override counts
+
 ### Changed
 
 ### Fixed